Improved documentation
