+++ events/messageCreate.js
@@ -18,7 +18,7 @@ module.exports = async (client, message) => {
   const settings = message.settings = getSettings(message.guild);
 
   //Ignore any channel which isn't #testing
-  //if (message.channel.name !== "testing") return;
+  if (message.channel.name !== "testing") return;
 
   // Checks if the bot was mentioned via regex, with no message after it,
   // returns the prefix. The reason why we used regex here instead of
