+++ modules/APIHelper.js
@@ -1,20 +1,15 @@
 const axios = require("axios");
 
 
-/*
-  .then(res => {
-      console.log(`statusCode: ${res.status}`);
-      return res;
-    })
-    .catch(error => {
-      console.error(error);
-    });
- */
-
 function getKey() {
   return process.env.API_KEY;
 }
 
+/**
+ * Get JSON object containing skills from database
+ * @return {Promise<AxiosResponse<any>>}
+ * @module APIHelper
+ */
 module.exports = function getSkills() {
   return axios
     .get(process.env.API_URL + "/v1/skills/", {
@@ -24,19 +19,15 @@ module.exports = function getSkills() {
     });
 };
 
-
+/**
+ * Get the JSON object for a skill with a given id from the database
+ * @param {string} id - The ObjectID of the skill to be requested
+ * @return {Promise<AxiosResponse<any>>}
+ * @module APIHelper
+ */
 module.exports = function getSkill(id) {
   return axios
     .get(process.env.API_URL + "/v1/skills/?id="+id)
-    .then(res => {
-      console.log(`statusCode: ${res.status}`);
-      return res;
-    });
-};
-
-module.exports = function resetTree(root, parent=null) {
-  return axios
-    .post(process.env.API_URL + "/v1/skills")
     .then(res => {
       console.log(`statusCode: ${res.status}`);
       return res;
