+++ modules/createSwipePanel.js
@@ -2,10 +2,13 @@ const {MessageActionRow, MessageButton, MessageEmbed} = require("discord.js");
 
 /**
  * Create swipeable panel
+ * Adds left/right embedded buttons to a discord message to navigate through the swipeable list
+ * List is cyclic and cycles back to the start
  * @param {Client} client - Discord bot client
  * @param {User} user - User who sent the message
  * @param {MessageEmbed} msg - Discord embedded message object
- * @param {Swipeable[]} list - List of swipeable skills
+ * @param {Swipeable[]} list - List of swipeable objects
+ * @module createSwipePanel
  */
 module.exports.createSwipePanel = function(client, user, msg, list) {
   let currentPage = 0;
